(#1327) refactor ispyb_callback classes as per PR discussion
